Initial import of Qt4 bindings, and for a Qt4 "Arthur" (QPainter) backend renderer.
The bindings are currently unstable - you can expect substantial change
in both source and binary interfaces.

The Arthur renderer currently does a reasonable job of rendering path
and fill, but the image rendering doesn't work (for reasons that aren't
clear to me) and text rendering doesn't use the right glyphs - it just
draws with the current font. There is a lot of work to do on this
too. Help is, of coure, welcome.