Use realloc/free instead of new/delete when resizing GooStrings
This allows for a large performance improvement when appending a large number
of characters to a GooString. This is especially helpful for TextOutputDev on
large PDFs. For example, the following code has the potential to be O(n) instead of
O(n²) with a good implementation of realloc.

while (n) {
  string.append(character);
  n--;
}