Refactor the PSConverter and add a PDFConverter (for PDF export/save).
Add an OutStream subclass to redirect the output to a QIODevice.
Split the PSConverter in a base class (BaseConverter) w/ d_ptr structure.
Add a new PDFConverter that inherit BaseConverter, and do the actual export
making use of the new QIODeviceOutStream.
The BaseConverter now handles automatically file or QIODevice output.