Update the field dict instead of the annot dict for non composed dicts
We were always updating the annot dictionary, adding or updating
the V field. While this is correct for composed dicts (annot + field)
it isn't for non composed ones since the annot dictionary doesn't
contain any V field because it's a field of the form field dictionary.
In these cases the form field is not correctly updated and if the
document is saved it will be wrong.