Major rework of text display in form fields code.
This adds support for UTF-16 strings as input, and will recode them to the
appropriate character set for the font.

The method Annot::layoutText packs a lot of functionality together, but I
couldn't see a good way to split it apart.  It will convert an input text
string from PDFDocEncoding or UTF-16 (encoding is autodetected) to the
appropriate encoding for the font which will be used to display the form
field.  At the same time, it can compute the width of the resulting string,
and optionally break the string at a specified width to aid in wrapping
multi-line text.  Text wrapping is integrated with encoding conversion
since wrapping is easiest with the original text (where spaces and newlines
are easily identified), but text widths can only be computed after
re-encoding.

Support for composite fonts is included, so long as those fonts use an
identity CMap, but this is still untested.  Support for more complex CMaps
is missing.

The rewrite also includes a minor improvement to the formatting of comb
fields; characters are now centered in the comb cells.