Allow grouped checkboxes to be selected individually.
When checkboxes are in a group, they ought to behave like a collection of
radio buttons.  However, when deciding whether to draw a checkbox as
selected or not, the checkbox-drawing code was looking up the value of the
V field in the form dictionary, which is shared among all checkboxes in the
group.  Thus, checkboxes would either all be on or off in unison.

Instead, look up the AS (appearance state) value in each checkbox's
annotation dictionary, so that different checkboxes can be drawn
differently.