Add in the initial part of the optional content support.
To see this work, compare ClarityOCGs.pdf with and
without this change.

Right now we only handle optional content using
XObjects. Optional content using Marked Content has
infrastructure, but is not implemented. That will be
quite invasive in Gfx, and I'm not confident enough
to do it this late in the process.