Fix regression in cairo output when transforming ctm
The commit:
ec01926e5a9dc16e200060497c43e79a1623698d "Avoid setting a singular ctm"
introduced a regression in the output when using a ctm.

The check for an invertable matrix had the side effect of inverting
the matrix. Instead, make a copy of the matrix before testing if it is
invertable.