Let compiler figure out CDECL
CDECL is a compiler specific macro to specify C calling convention.

On Windows, some functions must be forced to use this calling convention.
To do this, you can use the macro CDECL, which is defined in windef.h.

Poppler should not try to be smarter than the compiler and try to guess
the correct definition for CDECL. Instead it should rely on the
definitions in the windef.h file provided by the compiler.

On Unix, specifying a calling convention is not nescessary, so CDECL can
be an empty definition.