[glib] Several fixes in poppler_annot_get_color()
PopplerColor is compatible with GdkColor where every component is in the
range of 0 to 65535, we were returning values between 0 and 1 but
converted to guint16.
We were also showing a warning when AnnotColor::colorTransparent is
used, but we decided to return a PopplerColor * to be able to represent
transparent color returning NULL, so it's actually supported.