[Qt4] Do not try to resolve named destinations for GoTo links pointing to external documents.
In such cases, the named destination is a destination in the external document, so we would try to look up a destination which is not in the current document (thus the look up is unuseful).
It is task of the users of poppler-qt4 detect such situations, and resolve the named when necessary, using Document::linkDestination(QString).