[cairo] Fix drawImage() for non-1x1 images
Carlos noticed a nasty bug with converting drawImage() to use PAD + fill,
instead of NONE + paint. That is the image was being padded out far beyond
the correct output extents. The cause is that the caller pre-scales the
context for the image, so the output rectangle was many times the true
image size - obliterating large amounts of the page. The temporary fix is
to counter-act the scaling on the context. Longer term, after fixing all
painters to use PAD correctly, we need to review the callers to remove
unnecessary pre-scaling.