Don't use byte_lookup table when color space doesn't support getLine methods
For color spaces that don't implement getRGBLine or getGrayLine methods,
getRGB or getGray are called for every pixel, however we were allocating
the byte_lookup table and converting colors in those cases too. Instead
of falling back to generic methods in the base class, the new methods
useGetRGBLine and useGetGrayLine have been added to he base class, so that
when they are not suopported in the current color space byte_lookup table is
not used at all.

Fixes bug #11027.