[cairo] Apply a clip for masked drawImage
In order to use EXTEND_PAD with a mask, we need to apply a clip (to
constrain the image to the appropriate region). The complicating factor for
drawSoftImage() is that mask size is independent of the image size, so we
need to compute the intersection of the mask with in the image in userspace.