Don't render the color white in type 3 glyphs in the cairo backend
PDF allows the "g" operator in Type 3 charprocs but cairo user fonts
will render any stroke or fill in the font color.

As the only PDFs I've seen with "g" in the charprocs are only using
the gray values 0 or 1, a workaround is to disable strokes and fills
of the charproc when the gray level is > 0.5.