[annots] Use opacity when drawing Markup Annotations
According to the spec:

"The constant opacity value that shall be used in painting the
annotation. This value shall apply to all visible elements of the
annotation in its closed state (including its background and border) but
not to the pop-up window that appears when the annotation is opened.

The specified value shall not used if the annotation has an appearance
stream in that case, the appearance stream shall specify any
transparency.

If no explicit appearance stream is defined for the annotation, it may
be painted by implementation-dependent means that do not necessarily
conform to the PDF imaging model; in this case, the effect of this entry
is implementation-dependent as well."