[annots] Create appearance stream for Text Markup Annotations when not defined
Highlight annotations are an exception, we always ignore the
appearance stream to use our own, since for most of the documents the
appearance stream provided by the annotation is not enough. That's why
it's currently broken. This is what acroread does indeed and Leonard
Rosenthol recommended us to do the same.
See bug #23108.