Distinguish between columns and tables when selecting text
This commit add ability to detect tables in text by checking borders
of 4 neighbouring text blocks for arrangement (to the left, to the right,
center, ...). Detected border of whole table is then stored in ExMin, ExMax,
EyMin and EyMax of each block together with id of detected table. Sorting
of blocks is then performed on the these borders to be able to distinguish
tables from columns.
Pasting of selected text was modified so that tables are pasted correctly
(even with multi line cells).