[CMake/glib] add support for gobject-introspection
other than a module to find (using pkg-config) gobject-introspection, this adds a macro (gir_add_introspections) to add new introspections in a directory.
its working is much similar (even the variable names used) to the Makefile version provided by gobject-repository itself, with the notable difference of a disabled libtool in the g-ir-scanner invocation.
the only additional step needed is adding two custom targets, compiled by ALL, which have the gir and the typelib files as dependencies, to make them build with the usual `make all'.