Make the internal get_id() not fail because of null bytes in the ID.
Passing the const char* of the byte string to convert is not enough if its length must be checked,
as it might fail when the string of the ID contains null bytes.
Instead, pass the original GooString so its size is properly checked.

Also, remove an hardcoded 16 and make it dependent on pdfIdLength, as used elsewhere in get_id() function.