Poppler: Fix line selection, dont check y for Line selection
When a line is selected while the Y values of selection is
not in the y range of words, the render will render the
line with reverse color while the words does not show up.

This fix neglect the Y vaules to make sure that the words
also displayed when line is selected.